+++ src/popUp.css
@@ -8,7 +8,7 @@
     width: 100%;
     height: 100%;
     z-index: 5000;
-  }
+}
 
 .popUp-section {
     flex-direction: column;
@@ -16,10 +16,6 @@
     justify-content: center;
 }
 
-.serieImg {
-    
-}
-
 .xBttn {
     width: 50px;
-    }
+}
