+++ src/style.css
@@ -120,7 +120,7 @@ a {
   margin-bottom: 15%;
 }
 
-.pop-up,
+.popUp-bttn,
 .reserve {
   font-size: 0.8rem;
   margin: 3% auto;
